Refactor: front, back 분리
