+++ code/fastapi/Makefile
@@ -1,13 +1,7 @@
-run_black:
-	python3 -m black . -l 119
-
 run_server:
-	python3 -m app
+	python3 -m app.back
 
 run_client:
-	python3 -m streamlit run app/frontend.py --server.port 30001 --server.fileWatcherType none
+	python3 -m streamlit run app/front/main.py --server.port 30001 --server.fileWatcherType none
 
 run_app: run_server run_client
-
-run_assignment_tests:
-	poetry run pytest assignments/app_test.py
