+++ code/fastapi/app/README.md
@@ -1,10 +1,12 @@
-
+# 자동 채점 프로젝트
+```bash
+cd app
 ```
-├── README.md
-├── __main__.py
-├── confirm_button_hack.py
-├── frontend.py : Streamlit으로 작성한 프론트엔드
-├── main.py : backend.py 라고 생각해도 무방합니다(편의상 main.py라고 명명)
-├── main_with_comments.py : 주석이 포함된 라이브 코딩한 파일
-└── model.py : 딥러닝 모델
+## front-end
+```bash
+python3 -m streamlit run front/main.py --server.port 30001 --server.fileWatcherType none
+```
+## back-end
+```bash
+python3 -m back
 ```
