--- code/fastapi/app/__main__.py
@@ -1,4 +1,0 @@
-if __name__ == "__main__":
-    import uvicorn
-
-    uvicorn.run("app.main:app", host="127.0.0.1", port=8001, reload=True)
