+++ code/fastapi/app/log/predict_log.csv
@@ -1,0 +1,22 @@
+,label,predict,confidence
+0,0,3,0.9922208189964294
+1,0,2,0.9365431666374207
+2,0,1,0.9455953240394592
+3,0,4,0.9996737241744995
+4,0,5,0.8415554761886597
+5,0,2,0.9887407422065735
+6,0,1,0.9515353441238403
+7,0,4,0.9989004135131836
+8,0,4,0.9995344877243042
+9,0,2,0.9557877779006958
+10,0,1,0.9881749153137207
+11,0,4,0.999038577079773
+12,0,3,0.9647161364555359
+13,0,3,0.9936484694480896
+14,0,2,0.8489811420440674
+15,0,5,0.7949714064598083
+16,0,3,0.9602106809616089
+17,0,2,0.9920003414154053
+18,0,5,0.9843242168426514
+19,0,5,0.9019641280174255
+20,0,2,0.9487490653991699
