--- code/fastapi/app/model/cascade_mask_rcnn_convnext-t_fp16.py
@@ -1,107 +1,0 @@
-_base_ = [
-    './datasets/coco_detection.py',
-    './models/cascade_rcnn_r50_pafpn.py',
-    './schedules/schedule_adamw_1x_last.py',
-    './default_runtime.py',
-]
-
-# please install mmcls>=0.22.0
-# import mmcls.models to trigger register_module in mmcls
-custom_imports = dict(imports=['mmcls.models'], allow_failed_imports=False)
-checkpoint_file = 'https://download.openmmlab.com/mmclassification/v0/convnext/downstream/convnext-tiny_3rdparty_32xb128-noema_in1k_20220301-795e9634.pth'  # noqa
-
-model = dict(
-    backbone=dict(
-        _delete_=True,
-        type='mmcls.ConvNeXt',
-        arch='tiny',
-        out_indices=[0, 1, 2, 3],
-        drop_path_rate=0.4,
-        layer_scale_init_value=1.0,
-        gap_before_final_norm=False,
-        init_cfg=dict(
-            type='Pretrained', checkpoint=checkpoint_file,
-            prefix='backbone.')),
-    neck=dict(in_channels=[96, 192, 384, 768]),
-    roi_head=dict(bbox_head=[
-        dict(
-            type='ConvFCBBoxHead',
-            num_shared_convs=4,
-            num_shared_fcs=1,
-            in_channels=256,
-            conv_out_channels=256,
-            fc_out_channels=1024,
-            roi_feat_size=7,
-            num_classes=6,
-            bbox_coder=dict(
-                type='DeltaXYWHBBoxCoder',
-                target_means=[0., 0., 0., 0.],
-                target_stds=[0.1, 0.1, 0.2, 0.2]),
-            reg_class_agnostic=False,
-            reg_decoded_bbox=True,
-            norm_cfg=dict(type='BN', requires_grad=True),
-            loss_cls=dict(
-                type='CrossEntropyLoss', use_sigmoid=False, loss_weight=1.0),
-            loss_bbox=dict(type='GIoULoss', loss_weight=10.0)),
-        dict(
-            type='ConvFCBBoxHead',
-            num_shared_convs=4,
-            num_shared_fcs=1,
-            in_channels=256,
-            conv_out_channels=256,
-            fc_out_channels=1024,
-            roi_feat_size=7,
-            num_classes=6,
-            bbox_coder=dict(
-                type='DeltaXYWHBBoxCoder',
-                target_means=[0., 0., 0., 0.],
-                target_stds=[0.05, 0.05, 0.1, 0.1]),
-            reg_class_agnostic=False,
-            reg_decoded_bbox=True,
-            norm_cfg=dict(type='BN', requires_grad=True),
-            loss_cls=dict(
-                type='CrossEntropyLoss', use_sigmoid=False, loss_weight=1.0),
-            loss_bbox=dict(type='GIoULoss', loss_weight=10.0)),
-        dict(
-            type='ConvFCBBoxHead',
-            num_shared_convs=4,
-            num_shared_fcs=1,
-            in_channels=256,
-            conv_out_channels=256,
-            fc_out_channels=1024,
-            roi_feat_size=7,
-            num_classes=6,
-            bbox_coder=dict(
-                type='DeltaXYWHBBoxCoder',
-                target_means=[0., 0., 0., 0.],
-                target_stds=[0.033, 0.033, 0.067, 0.067]),
-            reg_class_agnostic=False,
-            reg_decoded_bbox=True,
-            norm_cfg=dict(type='BN', requires_grad=True),
-            loss_cls=dict(
-                type='CrossEntropyLoss', use_sigmoid=False, loss_weight=1.0),
-            loss_bbox=dict(type='GIoULoss', loss_weight=10.0))
-    ]))
-
-img_norm_cfg = dict(
-    mean=[123.675, 116.28, 103.53], std=[58.395, 57.12, 57.375], to_rgb=True)
-
-
-# optimizer = dict(
-#     _delete_=True,
-#     constructor='LearningRateDecayOptimizerConstructor',
-#     type='AdamW',
-#     lr=0.0002,
-#     betas=(0.9, 0.999),
-#     weight_decay=0.05,
-#     paramwise_cfg={
-#         'decay_rate': 0.7,
-#         'decay_type': 'layer_wise',
-#         'num_layers': 6
-#     })
-
-# lr_config = dict(warmup_iters=1000, step=[27, 33])
-# runner = dict(max_epochs=100)
-
-# you need to set mode='dynamic' if you are using pytorch<=1.5.0
-fp16 = dict(loss_scale=dict(init_scale=512))
