--- code/fastapi/app/model/default_runtime.py
@@ -1,30 +1,0 @@
-checkpoint_config = dict(interval=5)
-# yapf:disable
-log_config = dict(
-    interval=50,
-    hooks=[
-        dict(type='TextLoggerHook'),
-        # dict(type='TensorboardLoggerHook')
-    ])
-
-# yapf:enable
-custom_hooks = [dict(type="NumClassCheckHook")]
-
-dist_params = dict(backend="nccl")
-log_level = "INFO"
-load_from = None
-resume_from = None
-workflow = [("train", 1)]
-
-# disable opencv multithreading to avoid system being overloaded
-opencv_num_threads = 0
-# set multi-process start method as `fork` to speed up the training
-mp_start_method = "fork"
-
-evaluation = dict(interval=1, metric="bbox", classwise=True, save_best="bbox_mAP")
-
-# Default setting for scaling LR automatically
-#   - `enable` means enable scaling LR automatically
-#       or not by default.
-#   - `base_batch_size` = (8 GPUs) x (2 samples per GPU).
-auto_scale_lr = dict(base_batch_size=64)
