+++ code/fastapi/app/scoring_image/correct.png
