+++ code/fastapi/app/scoring_image/wrong.png
